+++ core_handlers.h
@@ -1,0 +1,108 @@
+/*
+  core_handlers.h - core event handler entry points and some reporting entry points
+
+  Part of grblHAL
+
+  Copyright (c) 2020-2021 Terje Io
+
+  Grbl is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  Grbl is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with Grbl.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+/*! \file
+    \brief core function pointers and data structures definitions.
+*/
+
+#pragma once
+
+#include "system.h"
+#include "stream.h"
+#include "settings.h"
+#include "report.h"
+
+/* TODO: add to grbl pointers so that a different formatting (xml, json etc) of reports may be implemented by driver?
+typedef struct {
+    status_code_t (*report_status_message)(status_code_t status_code);
+    alarm_code_t (*report_alarm_message)(alarm_code_t alarm_code);
+    message_code_t (*report_feedback_message)(message_code_t message_code);
+    void (*report_init_message)(void);
+    void (*report_grbl_help)(void);
+    void (*report_echo_line_received)(char *line);
+    void (*report_realtime_status)(void);
+    void (*report_probe_parameters)(void);
+    void (*report_ngc_parameters)(void);
+    void (*report_gcode_modes)(void);
+    void (*report_startup_line)(uint8_t n, char *line);
+    void (*report_execute_startup_message)(char *line, status_code_t status_code);
+} grbl_report_t;
+*/
+
+// Report entry points set by core at reset.
+
+typedef status_code_t (*status_message_ptr)(status_code_t status_code);
+typedef message_code_t (*feedback_message_ptr)(message_code_t message_code);
+
+typedef struct {
+    setting_output_ptr setting;
+    status_message_ptr status_message;
+    feedback_message_ptr feedback_message;
+} report_t;
+
+// Core event handler and other entry points.
+// Most of the event handlers defaults to NULL, a few is set up to call a dummy handler for simpler code.
+
+typedef void (*on_state_change_ptr)(sys_state_t state);
+typedef void (*on_probe_completed_ptr)(void);
+typedef void (*on_program_completed_ptr)(program_flow_t program_flow, bool check_mode);
+typedef void (*on_execute_realtime_ptr)(sys_state_t state);
+typedef void (*on_unknown_accessory_override_ptr)(uint8_t cmd);
+typedef bool (*on_unknown_realtime_cmd_ptr)(char c);
+typedef void (*on_report_options_ptr)(bool newopt);
+typedef void (*on_report_command_help_ptr)(void);
+typedef void (*on_global_settings_restore_ptr)(void);
+typedef setting_details_t *(*on_get_settings_ptr)(void); // NOTE: this must match the signature of the same definition in
+                                                         // the setting_details_t structure in settings.h!
+typedef void (*on_realtime_report_ptr)(stream_write_ptr stream_write, report_tracking_flags_t report);
+typedef void (*on_unknown_feedback_message_ptr)(stream_write_ptr stream_write);
+typedef bool (*on_laser_ppi_enable_ptr)(uint_fast16_t ppi, uint_fast16_t pulse_length);
+typedef status_code_t (*on_unknown_sys_command_ptr)(sys_state_t state, char *line); // return Status_Unhandled.
+typedef status_code_t (*on_user_command_ptr)(char *line);
+typedef sys_commands_t *(*on_get_commands_ptr)(void);
+
+typedef struct {
+    // report entry points set by core at reset.
+    report_t report;
+    // grbl core events - may be subscribed to by drivers or by the core.
+    on_state_change_ptr on_state_change;
+    on_probe_completed_ptr on_probe_completed;
+    on_program_completed_ptr on_program_completed;
+    on_execute_realtime_ptr on_execute_realtime;
+    on_unknown_accessory_override_ptr on_unknown_accessory_override;
+    on_report_options_ptr on_report_options;
+    on_report_command_help_ptr on_report_command_help;
+    on_global_settings_restore_ptr on_global_settings_restore;
+    on_get_settings_ptr on_get_settings;
+    on_realtime_report_ptr on_realtime_report;
+    on_unknown_feedback_message_ptr on_unknown_feedback_message;
+    on_unknown_realtime_cmd_ptr on_unknown_realtime_cmd;
+    on_unknown_sys_command_ptr on_unknown_sys_command; // return Status_Unhandled if not handled.
+    on_get_commands_ptr on_get_commands;
+    on_user_command_ptr on_user_command;
+    on_laser_ppi_enable_ptr on_laser_ppi_enable;
+    // core entry points - set up by core before driver_init() is called.
+    bool (*protocol_enqueue_gcode)(char *data);
+} grbl_t;
+
+extern grbl_t grbl;
+
+/*EOF*/